sanity_rs::config
Struct SanityConfig

pub struct SanityConfig {

    pub project_id: String,
    pub dataset: String,
    pub use_cdn: bool,
    pub token: Option<String>,
    pub perspective: ClientPerspective,
    pub api_host: Option<String>,
    pub api_version: Option<String>,
    pub proxy: Option<String>,
    pub request_tag_prefix: Option<String>,
    pub ignore_browser_token_warning: bool,
    pub with_credentials: bool,
    pub timeout: Option<u64>,
    pub max_retries: Option<u64>,
    pub retry_delay: Option<fn(_: u64) -> u64>,
    pub use_project_hostname: bool,

}

Configuration struct for the Sanity client.

This struct holds all the necessary configuration options for interacting with the Sanity API.

Fields
project_id: String

Your Sanity project ID. This is required.

dataset: String

Your Sanity dataset. This is required.

use_cdn: bool

Whether to use the Sanity CDN. Defaults to false.

token: Option<String>

Your Sanity API token. This is optional, but recommended for authenticated requests.

perspective: ClientPerspective

The perspective to use when querying Sanity. This determines which parts of your schema are visible. See the Sanity documentation for more details.

api_host: Option<String>

The host to use for the Sanity API. Defaults to the standard Sanity API host.

api_version: Option<String>

The API version to use. Defaults to the latest version.

proxy: Option<String>

Proxy settings to use for requests. Specifies the proxy URL as a string.

request_tag_prefix: Option<String>

Prefix for request tags. Useful for debugging and monitoring.

ignore_browser_token_warning: bool

Whether to ignore warnings about the browser token. Defaults to false.

with_credentials: bool

Whether to send cookies with requests. Defaults to false. This is primarily relevant for authenticated requests where cookies might be used for session management.

timeout: Option<u64>

Timeout for API requests in milliseconds.

max_retries: Option<u64>

Maximum number of retries for failed API requests.

retry_delay: Option<fn(_: u64) -> u64>

Function to calculate the delay between retries. The input is the retry number (starting at 0), and the output is the delay in milliseconds.

use_project_hostname: bool

Whether to use the project hostname in the API URL. Defaults to false.

Implementations

impl SanityConfig

pub fn new(project_id: String, dataset: String) -> Self